requirements.txt
mpi4py
matplotlib
numpy
google-generativeai